pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flipperforge"
version = "0.2.0"
description = "BadUSB payload workshop for Flipper Zero"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{name = "TiltedLunar123", email = "hilgendorfjude@gmail.com"}]
keywords = ["flipper-zero", "badusb", "duckyscript", "pentesting", "mitre-attack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyserial>=3.5",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
flipperforge = "flipperforge.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.setuptools.packages.find]
include = ["flipperforge*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
